requirements.txt
websockets>=12.0
pillow>=10.0.0
numpy>=1.26.0
